+++ src/plonemeeting/portal/core/browser/utils.py
@@ -10,6 +10,7 @@ from plonemeeting.portal.core.content.meeting import IMeeting
 from plonemeeting.portal.core.interfaces import IMeetingsFolder
 from plonemeeting.portal.core.interfaces import IPublicationsFolder
 from plonemeeting.portal.core.utils import get_term_title
+from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
 from Products.ZCatalog.interfaces import ICatalogBrain
 from zope.component import queryUtility
@@ -36,6 +37,9 @@ class UtilsView(BrowserView):
 
     def is_institution(self):
         return IInstitution.providedBy(self.context)
+
+    def is_in_institution(self):
+        return IInstitution.providedBy(api.portal.get_navigation_root(self.context))
 
     def is_meeting(self):
         return IMeeting.providedBy(self.context)
@@ -77,6 +81,16 @@ class UtilsView(BrowserView):
         return api.content.find(
             context=institution, object_provides=IPublicationsFolder.__identifier__
         )[0].getURL()
+
+    def get_settings_url(self):
+        institution = self.get_current_institution()
+        return f"{institution.absolute_url()}/@@manage-settings"
+
+    def show_settings_tab(self):
+        portal_membership = getToolByName(self.context, 'portal_membership')
+        if portal_membership.isAnonymousUser() or not self.is_in_institution():
+            return False
+        return portal_membership.checkPermission('Modify portal content', self.get_current_institution())
 
     @staticmethod
     def get_state(obj):