+++ CHANGES.rst
@@ -14,6 +14,11 @@ Changelog
   [aduchene]
 - DELIBE-24: Fix an issue where a decisions manager could not remove items from a meeting.
   [aduchene]
+- DELIBE-199: Users are now manageable by institution.
+  [aduchene]
+- DELIBE-199: Add new member and manager groups for each institution.
+  [aduchene]
+- DELIBE-199: New institution settings navigation tab and view.
 - Handle multiple possible cron paths in the Prometheus export view.
   [aduchene]